+++ src/main/java/luu/indepth/item/custom/tools/ModAxeItem.java
@@ -1,4 +1,4 @@
-package luu.indepth.item;
+package luu.indepth.item.custom.tools;
 
 import net.minecraft.item.AxeItem;
 import net.minecraft.item.ToolMaterial;
