+++ src/main/java/luu/indepth/item/custom/tools/ModHoe.java
@@ -1,6 +1,5 @@
-package luu.indepth.item;
+package luu.indepth.item.custom.tools;
 
-import net.minecraft.item.AxeItem;
 import net.minecraft.item.HoeItem;
 import net.minecraft.item.ToolMaterial;
